# fix(str): add invariant to avoid unexpected errors when parsing an invalid UTF8 string