--- a/src/Psl/Str/replace_ci.php
+++ b/src/Psl/Str/replace_ci.php
@@ -29,7 +29,7 @@ function replace_ci(string $haystack, string $needle, string $replacement, Encod
             'preg_split',
             static fn() => preg_split('{' . preg_quote($needle, '/') . '}iu', $haystack, -1),
         );
-    } catch (Regex\Exception\InvalidPatternException $error) {
+    } catch (Regex\Exception\RuntimeException | Regex\Exception\InvalidPatternException $error) {
         throw new Exception\InvalidArgumentException($error->getMessage(), previous: $error);
     }