--- a/src/Psl/Str/trim.php
+++ b/src/Psl/Str/trim.php
@@ -4,6 +4,8 @@ declare(strict_types=1);
 
 namespace Psl\Str;
 
+use Psl;
+
 use function preg_quote;
 use function preg_replace;
 
@@ -14,9 +16,13 @@ use function preg_replace;
  * be stripped: space, tab, newline, carriage return, NUL byte, vertical tab.
  *
  * @pure
+ *
+ * @throws Psl\Exception\InvariantViolationException if $string is not a valid UTF-8 string.
  */
 function trim(string $string, ?string $char_mask = null): string
 {
+    Psl\invariant(is_utf8($string), 'Expected $string to be a valid UTF-8 string.');
+
     $char_mask ??= " \t\n\r\0\x0B\x0C\u{A0}\u{FEFF}";
     $char_mask   = preg_quote($char_mask, null);