--- a/tests/unit/Str/ReplaceCiTest.php
+++ b/tests/unit/Str/ReplaceCiTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Psl\Tests\Unit\Str;
 
 use PHPUnit\Framework\TestCase;
+use Psl\Exception\InvariantViolationException;
 use Psl\Str;
 
 final class ReplaceCiTest extends TestCase
@@ -24,6 +25,38 @@ final class ReplaceCiTest extends TestCase
             ['Hello, World!', 'Hello, You!', 'You', 'World', ],
             ['مرحبا بكم', 'مرحبا سيف', 'سيف', 'بكم'],
             ['foo', 'foo', 'bar', 'baz'],
+        ];
+    }
+
+    /**
+     * @dataProvider provideBadUtf8Data
+     */
+    public function testBadUtf8(string $string, string $expectedException, string $expectedExceptionMessage): void
+    {
+        $this->expectException($expectedException);
+        $this->expectExceptionMessage($expectedExceptionMessage);
+
+        Str\replace_ci($string, $string, $string);
+    }
+
+    public function provideBadUtf8Data(): iterable
+    {
+        yield [
+            "\xc1\xbf",
+            InvariantViolationException::class,
+            'Expected $needle to be a valid UTF-8 string.',
+        ];
+
+        yield [
+            "\xe0\x81\xbf",
+            InvariantViolationException::class,
+            'Expected $needle to be a valid UTF-8 string.',
+        ];
+
+        yield [
+            "\xf0\x80\x81\xbf",
+            InvariantViolationException::class,
+            'Expected $needle to be a valid UTF-8 string.',
         ];
     }
 }