--- a/tests/unit/Str/ReplaceCiTest.php
+++ b/tests/unit/Str/ReplaceCiTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Psl\Tests\Unit\Str;
 
 use PHPUnit\Framework\TestCase;
-use Psl\Exception\InvariantViolationException;
 use Psl\Str;
 
 final class ReplaceCiTest extends TestCase
@@ -43,20 +42,20 @@ final class ReplaceCiTest extends TestCase
     {
         yield [
             "\xc1\xbf",
-            InvariantViolationException::class,
-            'Expected $needle to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Compilation failed: UTF-8 error: overlong 2-byte sequence at offset 0',
         ];
 
         yield [
             "\xe0\x81\xbf",
-            InvariantViolationException::class,
-            'Expected $needle to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Compilation failed: UTF-8 error: overlong 3-byte sequence at offset 0',
         ];
 
         yield [
             "\xf0\x80\x81\xbf",
-            InvariantViolationException::class,
-            'Expected $needle to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Compilation failed: UTF-8 error: overlong 4-byte sequence at offset 0',
         ];
     }
 }