--- a/tests/unit/Str/TrimLeftTest.php
+++ b/tests/unit/Str/TrimLeftTest.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace Psl\Tests\Unit\Str;
 
 use PHPUnit\Framework\TestCase;
+use Psl;
 use Psl\Str;
 
 final class TrimLeftTest extends TestCase
@@ -45,6 +46,38 @@ final class TrimLeftTest extends TestCase
                 "    Héllö     Wôrld\t!!!  \n",
                 ' ',
             ],
+        ];
+    }
+
+    /**
+     * @dataProvider provideBadUtf8Data
+     */
+    public function testBadUtf8(string $string, string $expectedException, string $expectedExceptionMessage): void
+    {
+        $this->expectException($expectedException);
+        $this->expectExceptionMessage($expectedExceptionMessage);
+
+        Str\trim_left($string);
+    }
+
+    public function provideBadUtf8Data(): iterable
+    {
+        yield [
+            "\xc1\xbf",
+            Psl\Exception\InvariantViolationException::class,
+            'Expected $string to be a valid UTF-8 string.',
+        ];
+
+        yield [
+            "\xe0\x81\xbf",
+            Psl\Exception\InvariantViolationException::class,
+            'Expected $string to be a valid UTF-8 string.',
+        ];
+
+        yield [
+            "\xf0\x80\x81\xbf",
+            Psl\Exception\InvariantViolationException::class,
+            'Expected $string to be a valid UTF-8 string.',
         ];
     }
 }