--- a/tests/unit/Str/TrimRightTest.php
+++ b/tests/unit/Str/TrimRightTest.php
@@ -5,7 +5,6 @@ declare(strict_types=1);
 namespace Psl\Tests\Unit\Str;
 
 use PHPUnit\Framework\TestCase;
-use Psl;
 use Psl\Str;
 
 final class TrimRightTest extends TestCase
@@ -74,20 +73,20 @@ final class TrimRightTest extends TestCase
     {
         yield [
             "\xc1\xbf",
-            Psl\Exception\InvariantViolationException::class,
-            'Expected $string to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Malformed UTF-8 characters, possibly incorrectly encoded',
         ];
 
         yield [
             "\xe0\x81\xbf",
-            Psl\Exception\InvariantViolationException::class,
-            'Expected $string to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Malformed UTF-8 characters, possibly incorrectly encoded',
         ];
 
         yield [
             "\xf0\x80\x81\xbf",
-            Psl\Exception\InvariantViolationException::class,
-            'Expected $string to be a valid UTF-8 string.',
+            Str\Exception\InvalidArgumentException::class,
+            'Malformed UTF-8 characters, possibly incorrectly encoded',
         ];
     }
 }