… docs/component/str.md