Fixes #409