@devnix 2023-06-17
I took the approach of adding an invariant checking that the problematic parameter is a valid utf8 before doing anything else. 

Maybe an error handling around `preg_replace` would feel safer but also this approach felt cleaner and more efficient to my eyes, let me know about anything I can improve or rewrite!