@veewee 2023-06-21
@devnix I took the liberty of performing some benchmarks here:

```php
#[Groups(['str'])]
final class Utf8Bench
{
    public function benchValidUtf8CheckThroughPreg(): void
    {
        preg_match('/./u', "HELLO");
    }

    public function benchValidUtf8CheckThroughEncodingAndPreg(): void
    {
        Psl\invariant(Str\is_utf8("HELLO"), 'Expected $string to be a valid UTF-8 string.');
        preg_match('/./u', "HELLO");
    }
}
```

The results are rather painful:

```
./vendor/bin/phpbench run --config config/phpbench.json --group=str --report=benchmark                                                                                                                  [13:37:38]
PHPBench (1.2.7) running benchmarks... #standwithukraine
with configuration file: /Users/verweto/Projects/github/azjezz/psl/config/phpbench.json
with PHP version 8.1.10, xdebug ✔, opcache ❌

\Psl\Tests\Benchmark\Str\Utf8Bench

    benchValidUtf8CheckThroughPreg..........R1 I2 - Mo0.700μs (±0.00%)
    benchValidUtf8CheckThroughEncodingAndPr.R4 I4 - Mo4.103μs (±1.18%)

Subjects: 2, Assertions: 0, Failures: 0, Errors: 0
Utf8Bench
=========

Average iteration times by variant

4.1μs     │   █
3.6μs     │   █
3.1μs     │   █
2.6μs     │   █
2.1μs     │   █
1.5μs     │   █
1.0μs     │ ▃ █
0.5μs     │ █ █
          └─────
            1 2

[█ <current>]

1: benchValidUtf8Check᠁ 2: benchValidUtf8Check᠁

Memory by variant

6.1mb     │ █ █
5.3mb     │ █ █
4.6mb     │ █ █
3.8mb     │ █ █
3.1mb     │ █ █
2.3mb     │ █ █
1.5mb     │ █ █
762.8kb   │ █ █
          └─────
            1 2

[█ <current>]

1: benchValidUtf8Check᠁ 2: benchValidUtf8Check᠁

<current>
+----------------------------------------------+---------+---------+--------+---------+
| subject                                      | memory  | mode    | rstdev | stdev   |
+----------------------------------------------+---------+---------+--------+---------+
| benchValidUtf8CheckThroughPreg ()            | 6.065mb | 0.700μs | ±0.00% | 0.000μs |
| benchValidUtf8CheckThroughEncodingAndPreg () | 6.102mb | 4.103μs | ±1.18% | 0.049μs |
+----------------------------------------------+---------+---------+--------+---------+
```

As you can see, this way is 8 times slower for `success`cases in comparison with dealing with the `null` return of `preg_x`.
It might not be the best of ideas - so I let @azjezz  decide on what to do with this :)