@azjezz 2023-06-21
I think using is_utf8 is not the right solution indeed, we could either manually check for null return, or use `Psl\Regex` if it doesn't introduce as much of an overhead.