@veewee 2023-06-21
@azjezz Psl\Regex is about 4.6x slower than regular `preg_`  - yet it is `4.600μs` so not really that big of a time :)

```php
#[Groups(['str'])]
final class Utf8Bench
{
    public function benchValidUtf8CheckThroughPreg(): void
    {
        if (null === preg_replace('/./u', '', "HELLO")) {
            throw new \RuntimeException('xxx');
        }
    }


    public function benchValidUtfWithPslRegex(): void
    {
        Regex\replace('HELLO', '/./u', '');
    }
}

```

```
PHPBench (1.2.7) running benchmarks... #standwithukraine
with configuration file: /Users/verweto/Projects/github/azjezz/psl/config/phpbench.json
with PHP version 8.1.10, xdebug ✔, opcache ❌

\Psl\Tests\Benchmark\Str\Utf8Bench

    benchValidUtf8CheckThroughPreg..........R1 I1 - Mo1.000μs (±0.00%)
    benchValidUtfWithPslRegex...............R1 I0 - Mo4.603μs (±1.06%)

Subjects: 2, Assertions: 0, Failures: 0, Errors: 0
Utf8Bench
=========

Average iteration times by variant

4.6μs     │   █
4.0μs     │   █
3.5μs     │   █
2.9μs     │   █
2.3μs     │   █
1.7μs     │   █
1.2μs     │ ▆ █
0.6μs     │ █ █
          └─────
            1 2

[█ <current>]

1: benchValidUtf8Check᠁ 2: benchValidUtfWithPs᠁

Memory by variant

6.1mb     │ █ █
5.3mb     │ █ █
4.5mb     │ █ █
3.8mb     │ █ █
3.0mb     │ █ █
2.3mb     │ █ █
1.5mb     │ █ █
758.0kb   │ █ █
          └─────
            1 2

[█ <current>]

1: benchValidUtf8Check᠁ 2: benchValidUtfWithPs᠁

<current>
+-----------------------------------+---------+---------+--------+---------+
| subject                           | memory  | mode    | rstdev | stdev   |
+-----------------------------------+---------+---------+--------+---------+
| benchValidUtf8CheckThroughPreg () | 6.064mb | 1.000μs | ±0.00% | 0.000μs |
| benchValidUtfWithPslRegex ()      | 6.064mb | 4.603μs | ±1.06% | 0.049μs |
+-----------------------------------+---------+---------+--------+---------+
```